+++ tests/assets/app/Router.php
@@ -2,32 +2,33 @@
 
 namespace App;
 
+use Symfony\Component\Routing\Route;
+
+
 /**
  * Router
  *
- * @uses \Phalcon\Mvc\Router\Group
+ * @uses Symfony\Component\Routing\RouteCollection
  * @copyright Akufen Atelier Creatif
  * @author Nicholas Charbonneau <nicholas@akufen.ca>
  */
-class Router extends \Phalcon\Mvc\Router\Group
+class Router extends \Symfony\Component\Routing\RouteCollection
 {
     /**
-     * Initialise the routing group
+     * Initialise the routing collection
      *
      * @return void
      */
-    public function initialize()
+    public function __construct()
     {
+        // Construct initial collection
+        parent::__construct();
+
         // Basic configuration
         $this->setPaths(array(
             'module' => 'app',
             'controller' => 'index',
             'namespace' => 'App\Controllers'
-        ));
-
-        $this->add('/test', array(
-            'controller' => 'testing',
-            'action' => 'static'
         ));
     }
 }