+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 0.1.5
     - Dispatch only published posts
     - Added ID getter and setter for Posts
+    - Fixed preventing Phalcon models to assign key values to properties
 
 0.1.4
     - Application handles only outside of admin panel