+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 0.1.5
 - Dispatch only published posts
 - Added ID getter and setter for Posts
+- Added permalink property to prevent multiple db acces
 - Fixed preventing Phalcon models to assign key values to properties
 - Fixed glitch when a property's value is null would throw an exception